fix(entities): widen tile friction to u16
Ten ground items in assets/items.yaml carry a tile_friction above 255
(up to 850). The Bevy client truncates friction through a u8, so on
those tiles it predicts a step 3-10x too short and its next walk is
refused. Widening the server side to u16 is half of aligning the two.

Pin the three shared answers against the client's matching test: the
formula is duplicated across two repositories and cannot share code, so
a divergence compiles cleanly on both sides and silently desyncs
movement.

+++ crates/server/src/entities/items.rs
@@ -54,7 +54,7 @@ pub enum ItemAttribute {
     Weight(u32),
     FloorChange(FloorChangeDirection),
     Inventory(InventorySlot),
-    TileFriction(u32),
+    TileFriction(u16),
     Action(ItemAction),
     MultiAction(ItemMultiAction),
     Decay { duration: Tick, decay_to: ItemId },
